# src/main/resources/static/docs/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/magazzino/prodotti":{"get":{"tags":["magazzino-rest-controller"],"summary":"Prodotti","description":"Ottieni solo i prodotti in giacenza","operationId":"prodottiInGiacenza","responses":{"200":{"description":"Lista dei prodotti","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Prodotto"}}}}}}}},"/api/magazzino/prodotti/{id}":{"get":{"tags":["magazzino-rest-controller"],"summary":"Prodotto","description":"Ottieni solo il prodotto che cerchi","operationId":"prdottoSpecifico","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Prodotto specifico","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Prodotto"}}}}}}},"/api/dashboard/fatturato":{"get":{"tags":["dashboard-rest-controller"],"summary":"Ottieni il fatturato mensile","description":"Calcola il totale delle fatture pagate nel mese corrente","operationId":"fatturatoMensile","responses":{"200":{"description":"Totale fatturato del mese","content":{"*/*":{"schema":{"type":"number"}}}}}}}},"components":{"schemas":{"Cliente":{"required":["cap","citta","email","indirizzo","provincia","ragioneSociale","telefono"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"ragioneSociale":{"type":"string"},"getpIvaCFiscale":{"maxLength":16,"minLength":11,"type":"string"},"email":{"type":"string"},"telefono":{"type":"string"},"indirizzo":{"type":"string"},"citta":{"type":"string"},"cap":{"type":"string"},"provincia":{"type":"string"},"fatture":{"type":"array","items":{"$ref":"#/components/schemas/Fattura"}},"ordine":{"type":"array","items":{"$ref":"#/components/schemas/Ordine"}}}},"DettaglioOrdine":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"quantita":{"type":"integer","format":"int32"},"ordine":{"$ref":"#/components/schemas/Ordine"},"prodotto":{"$ref":"#/components/schemas/Prodotto"}}},"Fattura":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dataInizioFattura":{"type":"string","format":"date"},"dataScadenzaFattura":{"type":"string","format":"date"},"iva":{"type":"number"},"cliente":{"$ref":"#/components/schemas/Cliente"},"statoIva":{"type":"string","enum":["IVA22","ESENTE","RIDOTTA"]},"statoFattura":{"type":"string","enum":["IN_ATTESA","PAGATO","SCADUTO"]},"importoTotale":{"type":"number"},"ordine":{"$ref":"#/components/schemas/Ordine"}}},"Ordine":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dettagli":{"type":"array","items":{"$ref":"#/components/schemas/DettaglioOrdine"}},"cliente":{"$ref":"#/components/schemas/Cliente"},"fattura":{"$ref":"#/components/schemas/Fattura"}}},"Prodotto":{"required":["quantitaDisponibile"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"prezzo":{"type":"number","format":"double"},"quantitaDisponibile":{"type":"integer","format":"int32"},"dettaglioOrdine":{"type":"array","items":{"$ref":"#/components/schemas/DettaglioOrdine"}},"upc":{"type":"string"}}}}}}